requirements.txt
python-dotenv
langchain
pandas
plotly
pymongo
requests
streamlit
ta
textblob
nixtla
setuptools